+++ crates/gmsol/src/cli/other.rs
@@ -97,6 +97,7 @@ impl Args {
         instruction_buffer: Option<InstructionBufferCtx<'_>>,
         serialize_only: Option<InstructionSerialization>,
         skip_preflight: bool,
+        priority_lamports: u64,
         max_transaction_size: Option<usize>,
     ) -> gmsol::Result<()> {
         match &self.command {
@@ -128,6 +129,7 @@ impl Args {
                     instruction_buffer,
                     serialize_only,
                     skip_preflight,
+                    Some(priority_lamports),
                     |signature| {
                         println!("{signature}");
                         Ok(())
@@ -189,6 +191,7 @@ impl Args {
                     instruction_buffer,
                     serialize_only,
                     skip_preflight,
+                    Some(priority_lamports),
                     |signature| {
                         println!("{signature}");
                         Ok(())
@@ -216,6 +219,7 @@ impl Args {
                     instruction_buffer,
                     serialize_only,
                     skip_preflight,
+                    Some(priority_lamports),
                     |signature| {
                         println!("{signature}");
                         Ok(())
@@ -240,6 +244,7 @@ impl Args {
                     instruction_buffer,
                     serialize_only,
                     skip_preflight,
+                    Some(priority_lamports),
                     |signature| {
                         tracing::info!("{signature}");
                         Ok(())
@@ -260,6 +265,7 @@ impl Args {
                     instruction_buffer,
                     serialize_only,
                     skip_preflight,
+                    Some(priority_lamports),
                     |signature| {
                         tracing::info!("{signature}");
                         Ok(())
@@ -279,6 +285,7 @@ impl Args {
                     instruction_buffer,
                     serialize_only,
                     skip_preflight,
+                    Some(priority_lamports),
                     |signature| {
                         tracing::info!("{signature}");
                         Ok(())
@@ -356,6 +363,7 @@ impl Args {
                     instruction_buffer,
                     serialize_only,
                     skip_preflight,
+                    Some(priority_lamports),
                 )
                 .await
             }