+++ crates/gmsol/src/cli/user.rs
@@ -36,6 +36,7 @@ impl Args {
         ctx: Option<InstructionBufferCtx<'_>>,
         serialize_only: Option<InstructionSerialization>,
         skip_preflight: bool,
+        priority_lamports: u64,
         max_transaction_size: Option<usize>,
     ) -> gmsol::Result<()> {
         let options = BundleOptions {
@@ -74,6 +75,7 @@ impl Args {
             ctx,
             serialize_only,
             skip_preflight,
+            Some(priority_lamports),
         )
         .await
     }